Fixed functionality for delete in custom card and name date card. forgot to bind the handleClick method. Added text to cart when it is empty

+++ client/pages/cart.jsx
@@ -112,18 +112,26 @@ export default class Cart extends React.Component {
   }
 
   render() {
-    return (
-      this.state.cartItems.map((product, index) => {
-        return (
+    if (this.state.cartItems.length === 0) {
+      return (
+        <div>
+          <p className='text-center'>You currently have no items in your cart.</p>
+        </div>
+      );
+    } else {
+      return (
+        this.state.cartItems.map((product, index) => {
+          return (
       <CartItem
           key={index}
           product={product}
           updateItem={this.updateItem}
           deleteItem={this.deleteItem}
       />
-        );
-      }
-      )
-    );
+          );
+        }
+        )
+      );
+    }
   }
 }
